Moved src/agent/schemas.py back into src/agent/agent_main.py. Debugged argument parsing in app.py
